refactor: tag 분리
